Create Employees_Earning_More.sql
This problem needs the use of 'JOIN' to complete. A 'JOIN' is a sql command that allows you to combine rows from two tables together based on related information or a matching column. For this problem, two tables have a column in common and you have to create a new table using that matching information from both columns(manager id and id).